+++ packages/riverpod_lint_flutter_test/test/lints/avoid_public_notifier_properties.g.dart
@@ -138,6 +138,11 @@ class GeneratedNotifierProvider
   }
 
   @override
+  (int,) get argument {
+    return (param,);
+  }
+
+  @override
   AutoDisposeNotifierProviderElement<GeneratedNotifier, int> createElement() {
     return _GeneratedNotifierProviderElement(this);
   }