+++ packages/riverpod_lint_flutter_test/test/lints/provider_parameters.g.dart
@@ -124,6 +124,13 @@ class GeneratorProvider extends Provider<int> {
   }
 
   @override
+  ({
+    Object? value,
+  }) get argument {
+    return (value: value,);
+  }
+
+  @override
   ProviderElement<int> createElement() {
     return _GeneratorProviderElement(this);
   }