+++ packages/riverpod_generator/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Annotating a provider with `@deprecated` and a few other annotations
   also annotate the generated code accordingly (thanks to @SunlightBro)
+- `provider.argument` is now a record of all arguments in a provider.
 
 ## 3.0.0-dev.3 - 2023-10-06